IODevice (HAL) Framework Phase 3
Rename PWM to Analogue (as it is just a digital to analogue converter, the PWM is done downstream).
Adjust sensor handling so the scan frequency and bounce filter are configured through constants in Sensors.h.
To save flash, don't create PCA9685 module on Nano or Uno unless specifically asked for in mySetup.h.
Complete driver for MCP23017 pending testing.